backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pymysql==1.1.0
cryptography==41.0.7
pdfplumber==0.10.3
PyPDF2==3.0.1
python-multipart==0.0.6
